fix migration

+++ system_management/app/Database/Migrations/2024-08-28-060341_AddIdToSalesTable.php
@@ -8,6 +8,9 @@ class AddIdToSalesTable extends Migration
 {
     public function up()
     {
+        if ($this->db->fieldExists('id', 'sales')) {
+            return;
+        }
         $this->db->query('ALTER TABLE sales DROP PRIMARY KEY');
 
         $this->db->query('ALTER TABLE sales ADD id INT(11) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY FIRST');
@@ -17,6 +20,9 @@ class AddIdToSalesTable extends Migration
     {
         $forge = \Config\Database::forge();
 
+        if (!$this->db->tableExists('sales')) {
+            return;
+        }
         if ($this->db->fieldExists('id', 'sales')) {
             // Drop the 'id' column if it exists
             $forge->dropColumn('sales', 'id');
